+++ app/src/main/java/host/stjin/anonaddy/ui/usernames/AddUsernameBottomDialogFragment.kt
@@ -2,7 +2,9 @@ package host.stjin.anonaddy.ui.usernames
 
 import android.app.Dialog
 import android.content.Context
+import android.os.Build
 import android.os.Bundle
+import android.text.Html
 import android.view.KeyEvent
 import android.view.LayoutInflater
 import android.view.View
@@ -18,7 +20,7 @@ import host.stjin.anonaddy_shared.NetworkHelper
 import kotlinx.coroutines.launch
 
 
-class AddUsernameBottomDialogFragment : BaseBottomSheetDialogFragment(), View.OnClickListener {
+class AddUsernameBottomDialogFragment(private val usernameLimit: Int) : BaseBottomSheetDialogFragment(), View.OnClickListener {
 
 
     private lateinit var listener: AddUsernameBottomDialogListener
@@ -49,15 +51,13 @@ class AddUsernameBottomDialogFragment : BaseBottomSheetDialogFragment(), View.On
         listener = parentFragment as AddUsernameBottomDialogListener
 
 
-        //TODO TEST IF THIS IS INDEED NOT REQUIRED ON ANDROID 14 AND OLDER
-
-//        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-//            binding.bsAddusernameUsernameDesc.text =
-//                (Html.fromHtml(requireContext().resources.getString(R.string.add_username_desc, usernameLimit), Html.FROM_HTML_MODE_COMPACT))
-//        } else {
-//            binding.bsAddusernameUsernameDesc.text =
-//                (Html.fromHtml(requireContext().resources.getString(R.string.add_username_desc, usernameLimit)))
-//        }
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+            binding.bsAddusernameUsernameDesc.text =
+                (Html.fromHtml(requireContext().resources.getString(R.string.add_username_desc, usernameLimit), Html.FROM_HTML_MODE_COMPACT))
+        } else {
+            binding.bsAddusernameUsernameDesc.text =
+                (Html.fromHtml(requireContext().resources.getString(R.string.add_username_desc, usernameLimit)))
+        }
 
         // 2. Setup a callback when the "Done" button is pressed on keyboard
         binding.bsAddusernameUsernameAddUsernameButton.setOnClickListener(this)
@@ -76,7 +76,7 @@ class AddUsernameBottomDialogFragment : BaseBottomSheetDialogFragment(), View.On
 
     companion object {
         fun newInstance(usernameLimit: Int): AddUsernameBottomDialogFragment {
-            return AddUsernameBottomDialogFragment()
+            return AddUsernameBottomDialogFragment(usernameLimit)
         }
     }